Move TokenPipeline into sub-directories
