+++ src/SimplifiedSearch/SearchPipelines/TokenPipelines/Components/ITokenPipelineComponent.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SimplifiedSearch.SearchPipelines
+namespace SimplifiedSearch.SearchPipelines.TokenPipelines.Components
 {
     internal interface ITokenPipelineComponent
     {
