+++ src/SimplifiedSearch/SearchPipelines/TokenPipelines/Components/TokenizeFilter.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SimplifiedSearch.SearchPipelines
+namespace SimplifiedSearch.SearchPipelines.TokenPipelines.Components
 {
     internal class TokenizeFilter : ITokenPipelineComponent
     {
